Java API - doesn't compile yet
