+++ akka-actor-typed/src/main/scala/akka/actor/typed/internal/routing/RoutingLogic.scala
@@ -68,5 +68,5 @@ object RoutingLogics {
     }
   }
 
-  def randomLogic[T]: RoutingLogic[T] = RandomLogic.asInstanceOf[RoutingLogic[T]]
+  def randomLogic[T](): RoutingLogic[T] = RandomLogic.asInstanceOf[RoutingLogic[T]]
 }
