+++ akka-actor-typed/src/main/scala/akka/actor/typed/javadsl/Routers.scala
@@ -2,7 +2,8 @@
  * Copyright (C) 2009-2019 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.actor.typed.scaladsl
+package akka.actor.typed.javadsl
+
 import akka.actor.typed.Behavior
 import akka.actor.typed.internal.routing.GroupRouterBuilder
 import akka.actor.typed.internal.routing.PoolRouterBuilder
@@ -49,14 +50,19 @@ object Routers {
 trait GroupRouter[T] extends Behavior[T] {
 
   /**
-   * Route messages by randomly selecting the routee from the available routees. This is the default for group routers.
-   * FIXME motivate when to use it
+   * Route messages by randomly selecting the routee from the available routees.
+   *
+   * Random routing makes it less likely that every `poolsize` message from a single producer ends up in the same
+   * mailbox of a slow actor
+   *
+   * This is the default for group routers.
    */
   def withRandomRouting(): GroupRouter[T]
 
   /**
    * Route messages by using round robin.
-   * FIXME motivate when to use it
+   *
+   * Round robin gives fair routing where every available routee gets the same amount of messages
    */
   def withRoundRobinRouting(): GroupRouter[T]
 
@@ -72,14 +78,18 @@ trait PoolRouter[T] extends Behavior[T] {
 
   /**
    * Route messages by randomly selecting the routee from the available routees.
-   * FIXME motivate when to use it
+   *
+   * Random routing makes it less likely that every `poolsize` message from a single producer ends up in the same
+   * mailbox of a slow actor.
    */
   def withRandomRouting(): PoolRouter[T]
 
   /**
    * Route messages through round robin, providing a fair distribution of messages across the routees.
+   *
+   * Round robin gives fair routing where every available routee gets the same amount of messages
+   *
    * This is the default for pool routers.
-   * FIXME motivate when to use it
    */
   def withRoundRobinRouting(): PoolRouter[T]
 }
